barcode-rest-api/requirements.txt
Flask
azure-cosmos
flask-cors
pytest